Begin creating templates
