--- .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,57 +1,0 @@
-name: Bug Report
-description: File a bug report
-labels: ["Type: Bug", "Status: Triage"]
-body:
-  - type: markdown
-    attributes:
-      value: >
-        Thanks for taking the time to fill out this bug report! Before submitting your issue, please make
-        sure you are using the latest version of the charm. If not, please switch to this image prior to 
-        posting your report to make sure it's not already solved.
-  - type: textarea
-    id: bug-description
-    attributes:
-      label: Bug Description
-      description: >
-        If applicable, add screenshots to help explain the problem you are facing.
-    validations:
-      required: true
-  - type: textarea
-    id: reproduction
-    attributes:
-      label: To Reproduce
-      description: >
-        Please provide a step-by-step instruction of how to reproduce the behavior.
-      placeholder: |
-        1. `juju deploy ...`
-        2. `juju relate ...`
-        3. `juju status --relations`
-    validations:
-      required: true
-  - type: textarea
-    id: environment
-    attributes:
-      label: Environment
-      description: >
-        We need to know a bit more about the context in which you run the charm.
-        - Are you running Juju locally, on lxd, in multipass or on some other platform?
-        - What track and channel you deployed the charm from (ie. `latest/edge` or similar).
-        - Version of any applicable components, like the juju snap, the model controller, lxd, microk8s, and/or multipass.
-    validations:
-      required: true
-  - type: textarea
-    id: logs
-    attributes:
-      label: Relevant log output
-      description: >
-        Please copy and paste any relevant log output. This will be automatically formatted into code, so no need for backticks.
-        Fetch the logs using `juju debug-log --replay` and `kubectl logs ...`. Additional details available in the juju docs 
-        at https://juju.is/docs/olm/juju-logs
-      render: shell
-    validations:
-      required: true
-  - type: textarea
-    id: additional-context
-    attributes:
-      label: Additional context
-
