--- .github/workflows/on_pull_request.yaml
@@ -1,8 +1,0 @@
-name: Tests
-on:
-  pull_request:
-
-jobs:
-  run-tests:
-    name: Run Tests
-    uses: ./.github/workflows/ci.yaml
