--- .github/workflows/on_push.yaml
@@ -1,12 +1,0 @@
-name: Tests
-on:
-  push:
-    branches:
-      - master
-      - main
-      - track/**
-
-jobs:
-  run-tests:
-    name: Run Tests
-    uses: ./.github/workflows/ci.yaml
