--- .jujuignore
@@ -1,3 +1,0 @@
-/venv
-*.py[cod]
-*.charm
