--- actions.yaml
@@ -1,16 +1,0 @@
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-#
-# TEMPLATE-TODO: change this example to suit your needs.
-# If you don't need actions, you can remove the file entirely.
-# It ties in to the example _on_fortune_action handler in src/charm.py
-#
-# Learn more about actions at: https://juju.is/docs/sdk/actions
-
-fortune:
-  description: Returns a pithy phrase.
-  params:
-    fail:
-      description: "Fail with this message"
-      type: string
-      default: ""
