+++ default/.gitignore
