+++ default/metadata.yaml
@@ -1,0 +1,41 @@
+# This file populates the Overview on Charmhub.
+# See https://juju.is/docs/some-url-to-be-determined/ for a checklist and guidance.
+
+# The charm package name, no spaces (required)
+# See https://juju.is/docs/sdk/naming#heading--naming-charms for guidance.
+name: {{ name }}
+
+# The following metadata are human-readable and will be published prominently on Charmhub.
+
+# (Recommended)
+display-name: Charm Template
+
+# (Required)
+summary: A very short one-line summary of the charm.
+
+description: |
+  A single sentence that says what the charm is, concisely and memorably.
+
+  A paragraph of one to three short sentences, that describe what the charm does.
+
+  A third paragraph that explains what need the charm meets.
+
+  Finally, a paragraph that describes whom the charm is useful for.
+
+# The containers and resources metadata apply to Kubernetes charms only.
+# Remove them if not required.
+
+# Your workload’s containers.
+containers:
+  httpbin:
+    resource: httpbin
+
+# This file populates the Resources tab on Charmhub.
+
+resources:
+  # An OCI image resource for each container listed above.
+  httpbin-image:
+    type: oci-image
+    description: OCI image for httpbin
+    # Included for simplicity in integration tests
+    upstream-source: kennethreitz/httpbin
