+++ default/requirements.txt
@@ -1,0 +1,1 @@
+ops >= 1.5.0
