+++ kubernetes/.gitignore
@@ -1,0 +1,8 @@
+venv/
+build/
+*.charm
+.tox/
+.coverage
+__pycache__/
+*.py[cod]
+.idea
