+++ kubernetes/charmcraft.yaml
@@ -1,0 +1,8 @@
+type: charm
+bases:
+  - build-on:
+      - name: "ubuntu"
+        channel: "20.04"
+    run-on:
+      - name: "ubuntu"
+        channel: "20.04"
