+++ kubernetes/pyproject.toml
@@ -1,6 +1,3 @@
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-
 # Testing tools configuration
 [tool.coverage.run]
 branch = true
@@ -35,5 +32,5 @@ per-file-ignores = ["tests/*:D100,D101,D102,D103,D104"]
 docstring-convention = "google"
 # Check for properly formatted copyright header in each file
 copyright-check = "True"
-copyright-author = "Canonical Ltd."
+copyright-author = "{{ author }}"
 copyright-regexp = "Copyright\\s\\d{4}([-,]\\d{4})*\\s+%(author)s"
