+++ kubernetes/src/charm.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python3
+# Copyright {{ year }} {{ author }}
+# See LICENSE file for licensing details.
+
+"""Charm the application."""
+
+import logging
+
+from ops.charm import CharmBase
+from ops.main import main
+from ops.model import ActiveStatus
+
+logger = logging.getLogger(__name__)
+
+
+class {{ class_name }}(CharmBase):
+    """Charm the application."""
+
+    def __init__(self, *args):
+        super().__init__(*args)
+        self.framework.observe(self.on.some_container_pebble_ready, self._on_pebble_ready)
+        self.framework.observe(self.on.config_changed, self._on_config_changed)
+
+    def _on_pebble_ready(self, event):
+        """Handle pebble-ready event"""
+        self.unit.status = ActiveStatus()
+
+if __name__ == "__main__":  # pragma: nocover
+    main({{ class_name }})
