+++ kubernetes/tests/unit/__init__.py
@@ -1,0 +1,5 @@
+# Copyright {{ year }} {{ author }}
+# See LICENSE file for licensing details.
+
+import ops.testing
+ops.testing.SIMULATE_CAN_CONNECT = True
