+++ kubernetes/tox.ini
@@ -1,4 +1,4 @@
-# Copyright 2022 Canonical Ltd.
+# Copyright {{ year }} {{ author }}
 # See LICENSE file for licensing details.
 
 [tox]
@@ -70,7 +70,7 @@ description = Run integration tests
 deps =
     pytest
     juju
-    pytest-operator>=0.13.0
+    pytest-operator
     -r{toxinidir}/requirements.txt
 commands =
     pytest -v --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs}
