--- requirements.txt
@@ -1,1 +1,0 @@
-ops >= 1.2.0
